% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sign.R
\name{my_sign}
\alias{my_sign}
\title{Determine your Astrological Sign}
\usage{
my_sign(birthmonth, birthday)
}
\arguments{
\item{birthmonth}{is a string Month}

\item{birthday}{is a numeric value 1-31}
}
\value{
astrological sign as string
}
\description{
Determine your Astrological Sign
}
\examples{
my_sign(March, 11)
"Pisces"
}
