% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich-option.R
\name{sandwich_ready}
\alias{sandwich_ready}
\title{Ready Sandwich}
\usage{
sandwich_ready(protein, topping, minute)
}
\arguments{
\item{protein}{can be a string of any main protein that you desire in your sanwich}

\item{topping}{can be any string containing a food}

\item{minute}{is a numeric for how many minutes}
}
\description{
A function to print out how long it is until a sandwich order is ready
}
\examples{

sandwich_ready("turkey", "lettuce", 15)
"Your turkey sandwich with lettuce and mayo will be ready in 15 minutes"
}
